Rewrite type exports to match runtime code
This commit refactors the index.d.ts file to match the actual exported
structure, which is important for compatibility with the tsconfig
`moduleResolution` option being set to `Node16` or `NodeNext`, which is
the recommended value for new projects.

As far as I can tell this doesn't break any compatibility with existing
projects that are still using the old value of `moduleResolution: Node`.
It might technically break the build if anyone was using named imports
of these values, but such code would fail to import at runtime anyway,
so this is actually desirable.